Move files out of the common package
The common package was used for sharing code between the model & schema
packages. The schema package is no more, so the common packages is not
needed anymore either.
