+++ src/main/kotlin/org/treeWare/model/action/GetVisitor.kt
@@ -1,6 +1,6 @@
 package org.treeWare.model.action
 
 import org.treeWare.model.operator.Leader1Follower2ModelVisitor
-import org.treeWare.common.traversal.TraversalAction
+import org.treeWare.model.operator.TraversalAction
 
 interface GetVisitor<MappingAux> : Leader1Follower2ModelVisitor<Unit, Unit, MappingAux, TraversalAction>
