+++ src/main/kotlin/org/treeWare/model/codec/JsonModelDecoder.kt
@@ -1,6 +1,6 @@
 package org.treeWare.model.codec
 
-import org.treeWare.common.codec.JsonWireFormatDecoder
+import org.treeWare.model.codec.decoder.JsonWireFormatDecoder
 import org.treeWare.model.codec.decoder.stateMachine.AuxDecodingStateMachine
 import org.treeWare.model.codec.decoder.stateMachine.DecodingStack
 import org.treeWare.model.codec.decoder.stateMachine.ModelDecodingStateMachine
