+++ src/main/kotlin/org/treeWare/model/codec/decoder/WireFormatDecoder.kt
@@ -1,5 +1,6 @@
-package org.treeWare.common.codec
+package org.treeWare.model.codec.decoder
 
+import org.treeWare.model.codec.decoder.stateMachine.DecodingStateMachine
 import java.io.Reader
 
 interface WireFormatDecoder {
