+++ src/main/kotlin/org/treeWare/model/codec/decoder/stateMachine/AssociationPathStateMachine.kt
@@ -1,6 +1,5 @@
 package org.treeWare.model.codec.decoder.stateMachine
 
-import org.treeWare.common.codec.AbstractDecodingStateMachine
 import org.treeWare.model.core.MutableEntityKeysModel
 
 class AssociationPathStateMachine<Aux>(
