+++ src/main/kotlin/org/treeWare/model/codec/decoder/stateMachine/AuxDecodingStateMachine.kt
@@ -1,7 +1,5 @@
 package org.treeWare.model.codec.decoder.stateMachine
 
-import org.treeWare.common.codec.DecodingStateMachine
-
 interface AuxDecodingStateMachine<Aux> : DecodingStateMachine {
     fun newAux()
     fun getAux(): Aux?
