+++ src/main/kotlin/org/treeWare/model/codec/decoder/stateMachine/BaseEntityStateMachine.kt
@@ -1,8 +1,6 @@
 package org.treeWare.model.codec.decoder.stateMachine
 
 import org.apache.logging.log4j.LogManager
-import org.treeWare.common.codec.AbstractDecodingStateMachine
-import org.treeWare.common.codec.SkipUnknownStateMachine
 import org.treeWare.metaModel.getFieldMeta
 import org.treeWare.metaModel.getFieldTypeMeta
 import org.treeWare.metaModel.getMetaName
