+++ src/main/kotlin/org/treeWare/model/codec/decoder/stateMachine/ModelDecodingStateMachine.kt
@@ -1,7 +1,6 @@
 package org.treeWare.model.codec.decoder.stateMachine
 
 import org.apache.logging.log4j.LogManager
-import org.treeWare.common.codec.DecodingStateMachine
 import org.treeWare.model.core.Model
 import org.treeWare.model.core.MutableModel
 import org.treeWare.model.core.Resolved
@@ -33,7 +32,7 @@ class ModelDecodingStateMachine<Aux>(
         reinitialize()
     }
 
-    fun reinitialize() {
+    private fun reinitialize() {
         stack.clear()
         stack.addFirst(modelStateMachine)
     }
