+++ src/main/kotlin/org/treeWare/model/codec/decoder/stateMachine/ModelStateMachine.kt
@@ -1,6 +1,5 @@
 package org.treeWare.model.codec.decoder.stateMachine
 
-import org.treeWare.common.codec.AbstractDecodingStateMachine
 import org.treeWare.model.core.Model
 import org.treeWare.model.core.MutableModel
 import org.treeWare.model.core.Resolved
