+++ src/main/kotlin/org/treeWare/model/codec/decoder/stateMachine/ValueDecodingStateMachine.kt
@@ -1,7 +1,5 @@
 package org.treeWare.model.codec.decoder.stateMachine
 
-import org.treeWare.common.codec.DecodingStateMachine
-
 interface ValueDecodingStateMachine<Aux> : DecodingStateMachine {
     fun setAux(aux: Aux)
 }
