+++ src/main/kotlin/org/treeWare/model/codec/encoder/AuxEncoder.kt
@@ -1,7 +1,5 @@
 package org.treeWare.model.codec.encoder
 
-import org.treeWare.common.codec.WireFormatEncoder
-
 interface AuxEncoder {
     fun encode(fieldName: String?, aux: Any?, wireFormatEncoder: WireFormatEncoder)
 }
