+++ src/main/kotlin/org/treeWare/model/operator/Leader1Follower0ModelForEach.kt
@@ -1,6 +1,5 @@
 package org.treeWare.model.operator
 
-import org.treeWare.common.traversal.TraversalAction
 import org.treeWare.model.core.*
 import org.treeWare.model.cursor.CursorMoveDirection
 import org.treeWare.model.cursor.LeaderModelCursor
